adding documentation

+++ metadataFunctions/aiFingerprintIngestV3.py
@@ -82,36 +82,39 @@ try:
   errorReport = ''
   outputFPFile = f"/opt/olympusat/resources/vionlabsReports/{cantemoItemId}_FP.json"
   
-  #------------------------------
-  # Making API call to Vionlabs to get fingerprints
-  headers = {
-    'Accept': 'application/json'
-  }
-  payload = {}
-  urlGetProfanitySegments = f"https://apis.prod.vionlabs.com/results/fingerprintplus/v1/{cantemoItemId}?&key=kt8cyimHXxUzFNGyhd7c7g"
-  # urlGetProfanitySegments = f"https://apis.prod.vionlabs.com/results/fingerprintplus/v1/OLT-003?&key=kt8cyimHXxUzFNGyhd7c7g"
-  httpApiResponse = requests.request("GET", urlGetProfanitySegments, headers=headers, data=payload)
-  httpApiResponse.raise_for_status()
-  responseJson = httpApiResponse.json()
-  #------------------------------
-  apiResponseJson = json.loads(httpApiResponse.text)
-  apiResponseJsonFormat = json.dumps(apiResponseJson, indent=2)
-  responseWriting = open(outputFPFile, "w")
-  responseWriting.write(apiResponseJsonFormat)
-  responseWriting.close()
+  if os.path.isfile(outputFPFile):
+    existingAnalysisRead = open(outputFPFile, "r")
+    responseJson = json.loads(existingAnalysisRead.read())
+    existingAnalysisRead.close()
+
+  else:
+    #------------------------------
+    # Making API call to Vionlabs to get fingerprints
+    headers = {
+      'Accept': 'application/json'
+    }
+    payload = {}
+    urlGetProfanitySegments = f"https://apis.prod.vionlabs.com/results/fingerprintplus/v1/{cantemoItemId}?&key=kt8cyimHXxUzFNGyhd7c7g"
+    # urlGetProfanitySegments = f"https://apis.prod.vionlabs.com/results/fingerprintplus/v1/OLT-003?&key=kt8cyimHXxUzFNGyhd7c7g"
+    httpApiResponse = requests.request("GET", urlGetProfanitySegments, headers=headers, data=payload)
+    httpApiResponse.raise_for_status()
+    responseJson = httpApiResponse.json()
+    #------------------------------
+    apiResponseJson = json.loads(httpApiResponse.text)
+    apiResponseJsonFormat = json.dumps(apiResponseJson, indent=2)
+    responseWriting = open(outputFPFile, "w")
+    responseWriting.write(apiResponseJsonFormat)
+    responseWriting.close()
 
   addingGenreLookup = "false"
   addingMoodLookup = "false"
   addingKeywordLookup = "false"
-  genreXML = ""
-  moodXML = ""
-  keywordXML = ""
   genreLookupXML = f"<SimpleMetadataDocument xmlns=\"http://xml.vidispine.com/schema/vidispine\">"
   moodLookupXML = f"<SimpleMetadataDocument xmlns=\"http://xml.vidispine.com/schema/vidispine\">"
   keywordLookupXML = f"<SimpleMetadataDocument xmlns=\"http://xml.vidispine.com/schema/vidispine\">"
-  genreXMLHeader = f"<MetadataDocument xmlns=\"http://xml.vidispine.com/schema/vidispine\"><group>Olympusat</group><timespan start=\"-INF\" end=\"+INF\"><field><name>oly_genreAnalysis</name>"
-  moodXMLHeader = f"<MetadataDocument xmlns=\"http://xml.vidispine.com/schema/vidispine\"><group>Olympusat</group><timespan start=\"-INF\" end=\"+INF\"><field><name>oly_moodAnalysis</name>"
-  keywordXML = f"<MetadataDocument xmlns=\"http://xml.vidispine.com/schema/vidispine\"><group>Olympusat</group><timespan start=\"-INF\" end=\"+INF\"><field><name>oly_keywordAnalysis</name>"
+  genreXML = f"<MetadataDocument xmlns=\"http://xml.vidispine.com/schema/vidispine\"><group>Olympusat</group><timespan start=\"-INF\" end=\"+INF\">"
+  moodXML = f"<MetadataDocument xmlns=\"http://xml.vidispine.com/schema/vidispine\"><group>Olympusat</group><timespan start=\"-INF\" end=\"+INF\">"
+  keywordXML = f"<MetadataDocument xmlns=\"http://xml.vidispine.com/schema/vidispine\"><group>Olympusat</group><timespan start=\"-INF\" end=\"+INF\">"
   #------------------------------
   # Parsing JSON and POST XML data
 
@@ -119,7 +122,7 @@ try:
   for individualGenre in genreList:
     genreLookupXML += f"<field><key>{individualGenre}</key><value>{individualGenre}</value></field>"
   for individualGenre in responseJson["genre"]:
-    genreXML = f"<value>{individualGenre}</value>" + genreXML
+    genreXML += f"<field><name>oly_genreAnalysis</name><value>{individualGenre}</value></field>"
     if individualGenre.lower() not in genreList:
       addingGenreLookup = "true"
       genreLookupXML += f"<field><key>{individualGenre}</key><value>{individualGenre}</value></field>"
@@ -129,7 +132,8 @@ try:
     genreLookupPayload = parsedGenreLookupXML.toprettyxml()
     # print(genreLookupPayload)
     createCantimoLookup("oly_genreAnalysis", genreLookupPayload)
-  genreXML = genreXMLHeader + genreXML + "</field></timespan></MetadataDocument>"
+    time.sleep(10)
+  genreXML += "</timespan></MetadataDocument>"
   parsedGenreXML = xml.dom.minidom.parseString(genreXML)
   genrePayload = parsedGenreXML.toprettyxml()
 
@@ -137,7 +141,7 @@ try:
   for individualMood in moodList:
     moodLookupXML += f"<field><key>{individualMood}</key><value>{individualMood}</value></field>"
   for individualMood in responseJson["mood_tag"]:
-    moodXML = f"<value>{individualMood}</value>" + moodXML
+    moodXML += f"<field><name>oly_moodAnalysis</name><value>{individualMood}</value></field>"
     if individualMood.lower() not in moodList:
       addingMoodLookup = "true"
       moodLookupXML += f"<field><key>{individualMood}</key><value>{individualMood}</value></field>"
@@ -147,7 +151,8 @@ try:
     moodLookupPayload = parsedMoodLookupXML.toprettyxml()
     # print(moodLookupPayload)
     createCantimoLookup("oly_moodAnalysis", moodLookupPayload)
-  moodXML = moodXMLHeader + moodXML + "</field></timespan></MetadataDocument>"
+    time.sleep(10)
+  moodXML += "</timespan></MetadataDocument>"
   parsedMoodXML = xml.dom.minidom.parseString(moodXML)
   moodPayload = parsedMoodXML.toprettyxml()
 
@@ -155,7 +160,7 @@ try:
   for individualKeyword in keywordList:
     keywordLookupXML += f"<field><key>{individualKeyword}</key><value>{individualKeyword}</value></field>"
   for individualKeyword in responseJson["keyword"]:
-    keywordXML += f"<value>{individualKeyword}</value>"
+    keywordXML += f"<field><name>oly_keywordAnalysis</name><value>{individualKeyword}</value></field>"
     if individualKeyword.lower() not in keywordList:
       addingKeywordLookup = "true"
       keywordLookupXML += f"<field><key>{individualKeyword}</key><value>{individualKeyword}</value></field>"
@@ -165,7 +170,8 @@ try:
     keywordLookupPayload = parsedKeywordLookupXML.toprettyxml()
     # print(keywordLookupPayload)
     createCantimoLookup("oly_keywordAnalysis", keywordLookupPayload)
-  keywordXML += "</field></timespan></MetadataDocument>"
+    time.sleep(10)
+  keywordXML += "</timespan></MetadataDocument>"
   parsedKeywordXML = xml.dom.minidom.parseString(keywordXML)
   keywordPayload = parsedKeywordXML.toprettyxml()
